+++ analytical/templatetags/clickmap.py
@@ -11,7 +11,7 @@ from django.template import Library, Node, TemplateSyntaxError
 from analytical.utils import is_internal_ip, disable_html, get_required_setting
 
 
-CLICKMAP_TRACKER_ID_RE = re.compile(r'^\d+$')
+CLICKMAP_TRACKER_ID_RE = re.compile(r'^\w+$')
 TRACKING_CODE = """
     <script type="text/javascript">
     var clickmapConfig = {tracker: '%(tracker_id)s', version:'2'};
@@ -47,7 +47,7 @@ class ClickmapNode(Node):
     def __init__(self):
         self.tracker_id = get_required_setting('CLICKMAP_TRACKER_ID',
                                                CLICKMAP_TRACKER_ID_RE,
-                                               "must be a (string containing) a number")
+                                               "must be an alphanumeric string")
 
     def render(self, context):
         html = TRACKING_CODE % {'tracker_id': self.tracker_id}